Splitting feature store types.ts
