+++ frontend/src/__mocks__/mockDataSources.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
 
-import { DataSource, DataSources } from '#~/pages/featureStore/types';
+import { DataSource, DataSources } from '#~/pages/featureStore/types/featureView';
 
 /* eslint-disable @typescript-eslint/naming-convention */
 export const mockDataSource_REQUEST_SOURCE = (partial?: Partial<DataSource>): DataSource => ({
